back/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
PyJWT==2.8.0
werkzeug==3.0.1
bcrypt==4.1.1
python-dateutil==2.8.2
